+++ demisto_sdk/commands/common/hook_validations/content_entity_validator.py
@@ -789,3 +789,13 @@ class ContentEntityValidator(BaseValidator):
                 return False
 
         return True
+
+    def is_line_ends_with_dot(self, text_to_test, arg_field):
+        line_with_missing_dot = ""
+        for arg in text_to_test.get(arg_field, []):
+            if not arg.get("description", "").endswith("."):
+                line_with_missing_dot += f"The argument {arg.get('name')} is missing '.' at the end of the sentence.\n"
+        for output in text_to_test.get("outputs", []):
+            if not output.get("description", "").endswith("."):
+                line_with_missing_dot += f"The context path {output.get('contextPath')} is missing '.' at the end of the sentence.\n"
+        return line_with_missing_dot